requirements.txt

dnspython
google-api-python-client
google-auth-oauthlib
oauth2client==3.0.0
lxml
psutil
motor
python-dotenv
python-magic
requests
pyrogram
hachoir
pytz
google-api-python-client
google-auth-httplib2
python-magic
telegraph
tenacity
uvloop
tgcrypto
beautifulsoup4
colorama
construct
cryptography
cssutils
ffmpy
future
HTMLParser
httpx
inquirer
isodate
natsort
nltk
pproxy
prettytable
protobuf==3.20.1
pycountry
pycryptodome
pycryptodomex
PyJWT
pymediainfo
pysrt
Requests
six
svtplay_dl
termcolor
titlecase
tldextract
toml
tqdm
Unidecode
urllib3
wcwidth
xmltodict
yarl
hpack
hyper
hyperframe
h2
expiringdict
cloudscraper